addedTimeLogicToApplication

+++ CodingTracker.carlosmorales125/CodingTracker/Models/CodingSession.cs
@@ -3,7 +3,7 @@ namespace CodingTracker.Models;
 public class CodingSession
 {
     public int Id { get; init; }
-    public int StartTime { get; init; }
-    public int EndTime { get; init; }
-    public int Duration => EndTime - StartTime;
+    public long StartTime { get; init; }
+    public long EndTime { get; init; }
+    public long Duration => EndTime - StartTime;
 }
